Organizing the so-called classes
