+++ src/utils.c
@@ -1,5 +1,3 @@
-#pragma once
-
 /*
  * This file is part of the luamotif-core distribution (https://github.com/digital-pet/luamotif-core).
  * Copyright (c) 2022 digital-pet.
@@ -17,4 +15,15 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>
  */
 
-int lm_NewRealize(lua_State* L);
+#include "include/common.h"
+#include "include/utils.h"
+
+char* gc_strdup(const char* s) {
+	int size = strlen(s) + 1;
+	char* p = GC_MALLOC(size);
+	if (p) {
+		memset(p, 0, size);
+		memcpy(p, s, size - 1);
+	}
+	return p;
+}
