+++ src/widgetfactory.c
@@ -15,32 +15,10 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>
  */
 
-#include <stdlib.h>
-#ifdef __linux__
-#include <bsd/string.h>
-#endif
-
-#include <X11/Xdefs.h>
-#include <Xm/XmAll.h>
-
-#include <gc.h>
-
-#include <lua.h>
-#include <lauxlib.h>
-
 #include "include/common.h"
+#include "include/utils.h"
 #include "include/luamotif.h"
-#include "include/callbacks.h"
 #include "include/widgetfactory.h"
-
-static char* gc_strdup(const char* s) {
-	int size = strlen(s) + 1;
-	char* p = GC_MALLOC(size);
-	if (p) {
-		memcpy(p, s, size - 1); //since GC_MALLOC always zeroes memory, we can copy size -1 and ensure there's a null terminator
-	}
-	return p;
-}
 
 int CreateManagedWidgetTree(lua_State* L, int parentObj, Widget wdgParent, char* pszWidgetName) {
 
@@ -52,6 +30,14 @@ int CreateManagedWidgetTree(lua_State* L, int parentObj, Widget wdgParent, char*
 	bool startManaged = true;
 
 	wdgWidget = NULL;
+
+	lua_pushstring(L, "__widget");
+	lua_rawget(L, -2);
+	if (lua_type(L, -1) == LUA_TLIGHTUSERDATA) {
+		lua_pop(L, 1);
+		return 1;
+	}
+	lua_pop(L, 1);
 
 	// TODO: If table already has widget, abort
 
@@ -135,146 +121,4 @@ int lm_NewRealize(lua_State* L) {
 	CreateManagedWidgetTree(L, 0, wdgToplevel, szWidgetName);
 	XtRealizeWidget(wdgToplevel);
 	return 0;
-}
-
-Widget ConstructGenericWidget(lua_State* L, int parentObj, Widget wdgParent, const char* pszWidgetName, MotifWidget WidgetFunction) {
-	_lua_stackguard_entry(L);
-
-	Arg aCreationArgs[MAXARGS];
-	Widget wdgWidget = NULL;
-	XmString* axmsValues;
-	int iArgCount = 0, iXmStringCount = 0, iLuaTableID;
-	char* pszKey, * pszValue;
-
-	struct cb_data* cbdCallback;
-
-	// Gather arguments by looping over the table
-	iLuaTableID = lua_gettop(L);
-	lua_pushnil(L);
-
-	axmsValues = (XmString*)GC_MALLOC(MAXARGS * sizeof(XmString));
-	if (axmsValues == NULL) {
-		luaL_error(L, "memory allocation failed");
-	}
-
-	while (lua_next(L, iLuaTableID) != 0) {
-		switch (lua_type(L, -2)) {
-		case LUA_TSTRING:
-			break;
-
-		default:
-			lua_pop(L, 1);
-			continue;
-		}
-
-		pszKey = gc_strdup(lua_tostring(L, -2));
-		if (pszKey == NULL) {
-			luaL_error(L, "memory allocation failed");
-		}
-
-		switch (lua_type(L, -1)) {
-		case LUA_TSTRING:
-			pszValue = gc_strdup(lua_tostring(L, -1));
-			if (pszValue == NULL) {
-				luaL_error(L, "memory allocation failed");
-			}
-
-
-			if (!strcmp(pszKey, "value") || !strcmp(pszKey, "title")) {
-				XtSetArg(aCreationArgs[iArgCount], pszKey, pszValue);
-			}
-			else {
-				axmsValues[iXmStringCount] = XmStringCreateLocalized(pszValue);
-				XtSetArg(aCreationArgs[iArgCount], pszKey, axmsValues[iXmStringCount]);
-				iXmStringCount++;
-			}
-			iArgCount++;
-			break;
-
-		case LUA_TNUMBER:
-			// TODO: this will fail because lua_tointeger doesn't guarantee it won't be garbage collected once it's off the stack.
-			XtSetArg(aCreationArgs[iArgCount], pszKey, lua_tointeger(L, -1));
-			iArgCount++;
-			break;
-
-		case LUA_TBOOLEAN:
-			// TODO: this will fail because lua_toboolean doesn't guarantee it won't be garbage collected once it's off the stack.
-			XtSetArg(aCreationArgs[iArgCount], pszKey, lua_toboolean(L, -1));
-			iArgCount++;
-			break;
-		}
-		lua_pop(L, 1);
-	}
-
-	// Create widget using function pointer with creation arguments
-
-	wdgWidget = (*WidgetFunction)(wdgParent, (String)pszWidgetName, aCreationArgs, iArgCount);
-
-	// Deallocate XmStrings which require deallocation
-	if (iXmStringCount > 0) {
-		while (--iXmStringCount >= 0) {
-			XmStringFree(axmsValues[iXmStringCount]);
-		}
-	}
-
-	if (wdgWidget == NULL) {
-		luaL_error(L, "failed to create widget");
-		// TODO: pass some sort of error up the chain
-	}
-
-	lua_pushlightuserdata(L, wdgWidget);
-	lua_setfield(L, -2, "__widget");
-	luaL_getmetatable(L, WIDGET_METATABLE);
-	lua_setmetatable(L, -2);
-
-	// Apply post-creation arguments
-	iLuaTableID = lua_gettop(L);
-	lua_pushnil(L);
-
-	while (lua_next(L, iLuaTableID) != 0) {
-		switch (lua_type(L, -2)) {
-		case LUA_TSTRING:
-			break;
-
-		default:
-			lua_pop(L, 1);
-			continue;
-		}
-		pszKey = gc_strdup(lua_tostring(L, -2));
-		// colors would be set here but meh I'll do that later.
-
-		switch (lua_type(L, -1)) {
-		case LUA_TFUNCTION:
-			cbdCallback = GC_malloc(sizeof(struct cb_data));
-			if (cbdCallback == NULL) {
-				luaL_error(L, "memory allocation failed");
-			}
-
-			cbdCallback->L = L;
-			lua_pushvalue(L, -1);
-
-			cbdCallback->ref = luaL_ref(L, LUA_REGISTRYINDEX);
-			lua_pushvalue(L, iLuaTableID);
-			
-			cbdCallback->obj = luaL_ref(L, LUA_REGISTRYINDEX);
-			cbdCallback->callback_name = gc_strdup(pszKey);
-
-			XtAddCallback(wdgWidget, pszKey, lm_Callback, cbdCallback);
-			XtAddCallback(wdgWidget, XmNdestroyCallback, lm_DestroyCallback, cbdCallback);
-			break;
-		}
-		lua_pop(L, 1);
-
-	}
-
-
-	// Return widget
-
-	_lua_stackguard_exit(L);
-	return wdgWidget;
-}
-
-Widget ConstructSimpleMenuBar(lua_State* L, int parentObj, Widget wdgParent, const char* pszWidgetName) {
-	// TODO: Everything
-	return;
 }
